requirements.txt
google-generativeai
python-dotenv
requests
streamlit
groq